pyproject.toml
[tool.poetry]
name = "wfc-python"
version = "0.1.0"
description = "Wave Function Collapse Python implimentation"
authors = ["E <e@pastel.codes>"]
readme = "README.md"
packages = [{ include = "wfc", from = "src" }]

[tool.poetry.scripts]
wfc = "wfc.main:main"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.2"
pillow = "^10.1.0"
pygame = "^2.5.2"
black = "^23.11.0"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
pycln = "^2.4.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pycln]
all = true

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
